Fixed cdash warning
Summary: make consistency  comparison

Test Plan: Compiles fine

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D5889

+++ branches/osgGeo-1.0/src/osgGeo/PolyLine.cpp
@@ -489,10 +489,10 @@ void PolyLineNode::removePrimitiveSet(int idx)
     if (idx<0)
 	return;
 
-    if (idx<_primitivesetModCount.size())
+    if (idx<(int)_primitivesetModCount.size())
 	_primitivesetModCount.erase(_primitivesetModCount.begin()+idx);
 
-    if (idx<_primitiveSets.size())
+    if (idx<(int)_primitiveSets.size())
 	_primitiveSets.erase(_primitiveSets.begin()+idx);
     
     touch();
